Refactor server and client
